Final Push for Jam
This has a bunch of major changes but using most of the system. The
prefabs OptionsCanvas were missing and could not be retrieved so they
need to be reestablished and re-uploaded WORKING, to git :+1:  Dialogue
is not written but system is there to replace the parts of the adlibs
and to use it for a higher level of lies and branches. The dialogue text
is still over top of the NPC text. Will fix another time along with
other missing things. FOr now this Jam has ended. Jam The Man 2018!
Thanks everyone.